Shifted indexing function for all dggs schemes from click annotated script to class
