+++ raster2dggs/indexers/rhprasterindexer.py
@@ -1,4 +1,20 @@
-from interfaces import RasterIndexer
+"""
+@author: ndemaio
+"""
+
+from numbers import Number
+from typing import Callable, Tuple, Union
+
+import rhppandas  # Necessary import despite lack of explicit use
+
+import pandas as pd
+import pyarrow as pa
+import xarray as xr
+import numpy as np
+
+import raster2dggs.constants as const
+
+from raster2dggs.interfaces import RasterIndexer
 
 class RHPRasterIndexer(RasterIndexer):
     '''
@@ -12,8 +28,35 @@ class RHPRasterIndexer(RasterIndexer):
             nodata: Number = np.nan,
             band_labels: Tuple[str] = None,
             ) -> pa.Table:
-        # TODO
-        pass
+        """
+        Index a raster window to rHEALPix.
+        Subsequent steps are necessary to resolve issues at the boundaries of windows.
+        If windows are very small, or in strips rather than blocks, processing may be slower
+        than necessary and the recommendation is to write different windows in the source raster.
+        """
+        sdf: pd.DataFrame = sdf.to_dataframe().drop(columns=["spatial_ref"]).reset_index()
+        subset: pd.DataFrame = sdf.dropna()
+        subset = subset[subset.value != nodata]
+        subset = pd.pivot_table(
+            subset, values=const.DEFAULT_NAME, index=["x", "y"], columns=["band"]
+        ).reset_index()
+        # Primary rHEALPix index
+        rhpindex = subset.rhp.geo_to_rhp(resolution, lat_col="y", lng_col="x").drop(
+            columns=["x", "y"]
+        )
+        # Secondary (parent) rHEALPix index, used later for partitioning
+        rhpindex = rhpindex.rhp.rhp_to_parent(parent_res).reset_index()
+        # Renaming columns to actual band labels
+        bands = sdf["band"].unique()
+        band_names = dict(zip(bands, map(lambda i: band_labels[i - 1], bands)))
+        for k, v in band_names.items():
+            if band_names[k] is None:
+                band_names[k] = str(bands[k - 1])
+            else:
+                band_names = band_names
+        rhpindex = rhpindex.rename(columns=band_names)
+        return pa.Table.from_pandas(rhpindex)
+
         
     def parent_groupby(
             self,
@@ -22,6 +65,14 @@ class RHPRasterIndexer(RasterIndexer):
             aggfunc: Union[str, Callable],
             decimals: int
             ) -> pd.DataFrame:
+        # TODO
+        pass
+        
+    def cell_to_children_size(
+            self,
+            cell,
+            desired_resolution: int
+            ) -> int:
         # TODO
         pass
