+++ raster2dggs/maidenhead.py
@@ -16,53 +16,53 @@ import raster2dggs.constants as const
 import raster2dggs.common as common
 from raster2dggs import __version__
 
-PAD_WIDTH = common.zero_padding("maidenhead")
+PAD_WIDTH = const.zero_padding("maidenhead")
 
 
-def _maidenheadfunc(
-    sdf: xr.DataArray,
-    level: int,
-    parent_level: int,
-    nodata: Number = np.nan,
-    band_labels: Tuple[str] = None,
-) -> pa.Table:
-    """
-    Index a raster window to Maidenhead.
-    Subsequent steps are necessary to resolve issues at the boundaries of windows.
-    If windows are very small, or in strips rather than blocks, processing may be slower
-    than necessary and the recommendation is to write different windows in the source raster.
-    """
-    sdf: pd.DataFrame = sdf.to_dataframe().drop(columns=["spatial_ref"]).reset_index()
-    subset: pd.DataFrame = sdf.dropna()
-    subset = subset[subset.value != nodata]
-    subset = pd.pivot_table(
-        subset, values=const.DEFAULT_NAME, index=["x", "y"], columns=["band"]
-    ).reset_index()
-    # Primary Maidenhead index
-    maidenhead = [
-        mh.to_maiden(lat, lon, level) for lat, lon in zip(subset["y"], subset["x"])
-    ]  # Vectorised
-    # Secondary (parent) Maidenhead index, used later for partitioning
-    maidenhead_parent = [
-        maidenhead_cell_to_parent(mh, parent_level) for mh in maidenhead
-    ]
-    subset = subset.drop(columns=["x", "y"])
-    subset[f"maidenhead_{level:0{PAD_WIDTH}d}"] = pd.Series(
-        maidenhead, index=subset.index
-    )
-    subset[f"maidenhead_{parent_level:0{PAD_WIDTH}d}"] = pd.Series(
-        maidenhead_parent, index=subset.index
-    )
-    # Rename bands
-    bands = sdf["band"].unique()
-    band_names = dict(zip(bands, map(lambda i: band_labels[i - 1], bands)))
-    for k, v in band_names.items():
-        if band_names[k] is None:
-            band_names[k] = str(bands[k - 1])
-        else:
-            band_names = band_names
-    subset = subset.rename(columns=band_names)
-    return pa.Table.from_pandas(subset)
+# def _maidenheadfunc(
+#     sdf: xr.DataArray,
+#     level: int,
+#     parent_level: int,
+#     nodata: Number = np.nan,
+#     band_labels: Tuple[str] = None,
+# ) -> pa.Table:
+#     """
+#     Index a raster window to Maidenhead.
+#     Subsequent steps are necessary to resolve issues at the boundaries of windows.
+#     If windows are very small, or in strips rather than blocks, processing may be slower
+#     than necessary and the recommendation is to write different windows in the source raster.
+#     """
+#     sdf: pd.DataFrame = sdf.to_dataframe().drop(columns=["spatial_ref"]).reset_index()
+#     subset: pd.DataFrame = sdf.dropna()
+#     subset = subset[subset.value != nodata]
+#     subset = pd.pivot_table(
+#         subset, values=const.DEFAULT_NAME, index=["x", "y"], columns=["band"]
+#     ).reset_index()
+#     # Primary Maidenhead index
+#     maidenhead = [
+#         mh.to_maiden(lat, lon, level) for lat, lon in zip(subset["y"], subset["x"])
+#     ]  # Vectorised
+#     # Secondary (parent) Maidenhead index, used later for partitioning
+#     maidenhead_parent = [
+#         maidenhead_cell_to_parent(mh, parent_level) for mh in maidenhead
+#     ]
+#     subset = subset.drop(columns=["x", "y"])
+#     subset[f"maidenhead_{level:0{PAD_WIDTH}d}"] = pd.Series(
+#         maidenhead, index=subset.index
+#     )
+#     subset[f"maidenhead_{parent_level:0{PAD_WIDTH}d}"] = pd.Series(
+#         maidenhead_parent, index=subset.index
+#     )
+#     # Rename bands
+#     bands = sdf["band"].unique()
+#     band_names = dict(zip(bands, map(lambda i: band_labels[i - 1], bands)))
+#     for k, v in band_names.items():
+#         if band_names[k] is None:
+#             band_names[k] = str(bands[k - 1])
+#         else:
+#             band_names = band_names
+#     subset = subset.rename(columns=band_names)
+#     return pa.Table.from_pandas(subset)
 
 
 def _maidenhead_parent_groupby(
@@ -87,11 +87,11 @@ def _maidenhead_parent_groupby(
         )
 
 
-def maidenhead_cell_to_parent(cell: str, parent_level: int) -> str:
-    """
-    Returns cell parent at some offset level.
-    """
-    return cell[: parent_level * 2]
+# def maidenhead_cell_to_parent(cell: str, parent_level: int) -> str:
+#     """
+#     Returns cell parent at some offset level.
+#     """
+#     return cell[: parent_level * 2]
 
 
 def maidenhead_cell_to_children_size(cell: str, desired_level: int) -> int:
@@ -262,13 +262,14 @@ def maidenhead(
         warp_mem_limit,
         resampling,
         overwrite,
+        compact,
     )
 
     common.initial_index(
         "maidenhead",
-        _maidenheadfunc,
-        _maidenhead_parent_groupby,
-        _maidenhead_compaction if compact else None,
+        #_maidenheadfunc, #TODO: Call from interface
+        _maidenhead_parent_groupby, #TODO: Call from interface
+        _maidenhead_compaction if compact else None, #TODO: Call from interface
         raster_input,
         output_directory,
         int(resolution),
