+++ tests/test_runthrough.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 @author: ndemaio
 """
